feat: separate registration page with confirm password
